requirements.txt
Pillow==10.0.0
matplotlib==3.7.2
shapely==2.0.1
scikit-image==0.21.0